moves routes and db functions out of main.go
