+++ lang/storj-node-go/.gitignore
@@ -1,2 +1,1 @@
-vendor/
-
+*.swp
